+++ carp_audio_package/test/carp_audio_package_test.dart
@@ -14,8 +14,10 @@ void main() {
 
     study = Study("1234", "bardram", name: "bardram study")
       ..dataEndPoint = DataEndPoint(DataEndPointTypes.PRINT)
-      ..addTriggerTask(ImmediateTrigger(),
-          Task(name: 'Task #1')..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
+      ..addTriggerTask(
+          ImmediateTrigger(),
+          AutomaticTask(name: 'Task #1')
+            ..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
   });
 
   test('Study -> JSON', () async {