+++ carp_mobile_sensing/example/lib/example.dart
@@ -23,7 +23,7 @@ void example() async {
   // careful - these sensors generate a lot of data!
   study.addTriggerTask(
       DelayedTrigger(delay: 1000), // delay sampling for one second
-      Task(name: 'Sensor Task')
+      AutomaticTask(name: 'Sensor Task')
         ..addMeasure(PeriodicMeasure(
           MeasureType(NameSpace.CARP, SensorSamplingPackage.ACCELEROMETER),
           frequency: 10 * 1000, // sample every 10 secs
@@ -37,7 +37,7 @@ void example() async {
 
   study.addTriggerTask(
       PeriodicTrigger(period: 24 * 60 * 60 * 1000), // trigger sampling once pr. day
-      Task(name: 'Task collecting a list of all installed apps')
+      AutomaticTask(name: 'Task collecting a list of all installed apps')
         ..addMeasure(Measure(MeasureType(NameSpace.CARP, AppsSamplingPackage.APPS))));
 
   // creating measure variable to be used later
@@ -47,7 +47,7 @@ void example() async {
     frequency: 11 * 1000,
     duration: 700,
   );
-  study.addTriggerTask(ImmediateTrigger(), Task(name: 'Light')..addMeasure(lightMeasure));
+  study.addTriggerTask(ImmediateTrigger(), AutomaticTask(name: 'Light')..addMeasure(lightMeasure));
 
   // Create a Study Controller that can manage this study.
   StudyController controller = StudyController(study);
@@ -123,7 +123,7 @@ void samplingSchemaExample() async {
     ..name = 'CARP Mobile Sensing - default configuration'
     ..dataEndPoint = DataEndPoint(DataEndPointTypes.PRINT)
     ..addTriggerTask(ImmediateTrigger(),
-        Task()..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
+        AutomaticTask()..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
 
   Study study = Study("2", 'user@cachet.dk',
       name: 'A outdoor activity study',
@@ -136,7 +136,7 @@ void samplingSchemaExample() async {
   // adding a set of specific measures from the `common` sampling schema to one overall task
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task()
+      AutomaticTask()
         ..measures = SamplingSchema.common().getMeasureList(
           namespace: NameSpace.CARP,
           types: [
@@ -150,7 +150,7 @@ void samplingSchemaExample() async {
   // adding a set of specific measures from the `common` sampling schema to one overall task
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task(name: 'Sensing Task #1')
+      AutomaticTask(name: 'Sensing Task #1')
         ..measures = SamplingSchema.common().getMeasureList(
           namespace: NameSpace.CARP,
           types: [
@@ -162,7 +162,7 @@ void samplingSchemaExample() async {
 
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task(name: 'One Common Sensing Task')
+      AutomaticTask(name: 'One Common Sensing Task')
         ..measures = SamplingSchema.common().getMeasureList(
           namespace: NameSpace.CARP,
           types: [
@@ -174,12 +174,13 @@ void samplingSchemaExample() async {
         ));
 
   // adding all measure from the activity schema to one overall 'sensing' task
-  study.addTriggerTask(ImmediateTrigger(), Task(name: 'Sensing Task')..measures = activitySchema.measures.values);
+  study.addTriggerTask(
+      ImmediateTrigger(), AutomaticTask(name: 'Sensing Task')..measures = activitySchema.measures.values);
 
   // adding the measures to two separate tasks, while also adding a new light measure to the 2nd task
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task(name: 'Activity Sensing Task #1')
+      AutomaticTask(name: 'Activity Sensing Task #1')
         ..measures = activitySchema.getMeasureList(
           namespace: NameSpace.CARP,
           types: [
@@ -191,7 +192,7 @@ void samplingSchemaExample() async {
 
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task(name: 'Phone Sensing Task #2')
+      AutomaticTask(name: 'Phone Sensing Task #2')
         ..measures = activitySchema.getMeasureList(
           namespace: NameSpace.CARP,
           types: [
@@ -234,12 +235,12 @@ void example_2() {
         ..zip = true
         ..encrypt = false)
     ..addTriggerTask(ImmediateTrigger(),
-        Task()..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
+        AutomaticTask()..measures = SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList());
 
   // adding a set of specific measures from the `common` sampling schema to one no-name task
   study.addTriggerTask(
       ImmediateTrigger(),
-      Task()
+      AutomaticTask()
         ..measures = SamplingSchema.common().getMeasureList(
           types: [SensorSamplingPackage.PEDOMETER, DeviceSamplingPackage.SCREEN],
           namespace: NameSpace.CARP,