+++ carp_mobile_sensing/lib/domain/triggers.dart
@@ -13,17 +13,21 @@ part of domain;
 /// be time-bound, based on data streams, initiated by a user of the platform,
 /// or a combination of these.
 ///
-/// Sub-classes of [Trigger] implements the specific behavior / timing of a trigger.
-/// Note that you should **not** use/instantiate this [Trigger] base class, but only its subclasses.
+/// The [Trigger] class is abstract. Use sub-classes of [Trigger] implements
+/// the specific behavior / timing of a trigger.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
-class Trigger extends Serializable {
+abstract class Trigger extends Serializable {
+  /// A unique id of this trigger.
+  /// Is used when storing data to know what triggered the data collection.
+  String triggerId;
+
   /// The list of [Task]s in this [Trigger].
   List<Task> tasks = new List<Task>();
 
   /// Add a [Task] to this [Trigger]
   void addTask(Task task) => tasks.add(task);
 
-  Trigger() : super();
+  Trigger({this.triggerId}) : super();
 
   static Function get fromJsonFunction => _$TriggerFromJson;
   factory Trigger.fromJson(Map<String, dynamic> json) =>
@@ -34,7 +38,7 @@ class Trigger extends Serializable {
 /// A trigger that starts sampling immediately and never stops.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class ImmediateTrigger extends Trigger {
-  ImmediateTrigger() : super();
+  ImmediateTrigger([String triggerId]) : super(triggerId: triggerId);
 
   static Function get fromJsonFunction => _$ImmediateTriggerFromJson;
   factory ImmediateTrigger.fromJson(Map<String, dynamic> json) =>
@@ -48,7 +52,7 @@ class ImmediateTrigger extends Trigger {
 /// Note that sampling continues until it is manually paused.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class ManualTrigger extends Trigger {
-  ManualTrigger() : super();
+  ManualTrigger({String triggerId}) : super(triggerId: triggerId);
 
   @JsonKey(ignore: true)
   ManualTriggerExecutor executor;
@@ -81,7 +85,7 @@ class DelayedTrigger extends Trigger {
   /// Delay in milliseconds.
   int delay;
 
-  DelayedTrigger({this.delay = 0}) : super();
+  DelayedTrigger({String triggerId, this.delay = 0}) : super(triggerId: triggerId);
 
   static Function get fromJsonFunction => _$DelayedTriggerFromJson;
   factory DelayedTrigger.fromJson(Map<String, dynamic> json) =>
@@ -103,7 +107,7 @@ class PeriodicTrigger extends Trigger {
   /// The duration (until paused) of the the sampling in milliseconds.
   int duration;
 
-  PeriodicTrigger({this.period = 60 * 1000, this.duration = 1000}) : super();
+  PeriodicTrigger({String triggerId, this.period = 60 * 1000, this.duration = 1000}) : super(triggerId: triggerId);
 
   static Function get fromJsonFunction => _$PeriodicTriggerFromJson;
   factory PeriodicTrigger.fromJson(Map<String, dynamic> json) =>
@@ -121,7 +125,7 @@ class ScheduledTrigger extends Trigger {
   /// If null, the sampling is never stopped (i.e., runs forever).
   int duration;
 
-  ScheduledTrigger({@required this.schedule, this.duration}) : super();
+  ScheduledTrigger({String triggerId, @required this.schedule, this.duration}) : super(triggerId: triggerId);
 
   static Function get fromJsonFunction => _$ScheduledTriggerFromJson;
   factory ScheduledTrigger.fromJson(Map<String, dynamic> json) =>
@@ -255,7 +259,8 @@ class RecurrentScheduledTrigger extends PeriodicTrigger {
 //  int monthOfYear;
 
   RecurrentScheduledTrigger(
-      {@required this.type,
+      {String triggerId,
+      @required this.type,
       @required this.time,
       this.end,
       this.separationCount = 0,
@@ -265,7 +270,7 @@ class RecurrentScheduledTrigger extends PeriodicTrigger {
       //this.dayOfMonth,
       int duration = 1000})
       : assert(duration != null),
-        super(duration: duration);
+        super(triggerId: triggerId, duration: duration);
 
   DateTime get firstOccurrence {
     DateTime _firstOccurrence;
@@ -314,7 +319,12 @@ class RecurrentScheduledTrigger extends PeriodicTrigger {
 /// For example, if [measureType] is `carp.geofence` the [resumeCondition] can be `{'DTU','ENTER'}`
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class SamplingEventTrigger extends Trigger {
-  SamplingEventTrigger({this.measureType, this.resumeCondition, this.pauseCondition}) : super();
+  SamplingEventTrigger({
+    String triggerId,
+    this.measureType,
+    this.resumeCondition,
+    this.pauseCondition,
+  }) : super(triggerId: triggerId);
 
   /// The [MeasureType] of the event to look for.
   ///
@@ -349,14 +359,19 @@ typedef EventConditionEvaluator = bool Function(Datum datum);
 /// a application-specific [condition] is meet.
 ///
 /// In contrast to other [Trigger]s, this trigger cannot be de/serialized from/to JSON.
-/// This implies that if can not be retrieved as part of a [Study] from a [StudyManager]
+/// This implies that it can not be retrieved as part of a [Study] from a [StudyManager]
 /// since it relies on specifying a Dart-specific function as the [EventConditionEvaluator]
 /// method. Hence, this trigger is mostly useful when creating a [Study] directly in the app
 /// using Dart code.
 ///
 /// If you need to de/serialize an event trigger, use the [SamplingEventTrigger] instead.
 class ConditionalSamplingEventTrigger extends Trigger {
-  ConditionalSamplingEventTrigger({this.measureType, this.resumeCondition, this.pauseCondition}) : super();
+  ConditionalSamplingEventTrigger({
+    String triggerId,
+    this.measureType,
+    this.resumeCondition,
+    this.pauseCondition,
+  }) : super(triggerId: triggerId);
 
   /// The [MeasureType] of the event to look for.
   MeasureType measureType;