+++ carp_mobile_sensing_app/lib/src/sensing/sensing.dart
@@ -314,7 +314,7 @@ class StudyMock implements StudyManager {
             // AUDIO and NOISE cannot be used in the same study since they conflict in using the microphone...
             ..addTriggerTask(
                 PeriodicTrigger(period: 1 * 60 * 1000, duration: 5 * 1000),
-                Task('Audio')
+                AutomaticTask(name: 'Audio')
                   ..measures.add(AudioMeasure(
                     MeasureType(NameSpace.CARP, AudioSamplingPackage.AUDIO),
                     name: "Audio Recording",