+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.1
+* Added `trigger_id` to the `Trigger` class.
+* Made `Task` abstract rather than deprecated.
+
 ## 0.7.0
 * **BREAKING** - this release break some parts of the previous API.
 * Extended Task model -- introduction of a `AutomaticTask` and `AppTask`.